backend/requirements.txt
Flask>=2.3.0,<3.0.0
flask-cors>=4.0.0,<5.0.0
pyproj>=3.4.0,<4.0.0
python-dotenv>=1.0.0,<2.0.0
geopandas>=0.13.0
werkzeug>=2.3.0
rasterio>=1.3.0
gunicorn>=21.2.0
pandas>=2.0.0
